refactor: move SnowFlake connector to a dedicate file
